+++ options.js
@@ -1,17 +1,3 @@
-// import functions from storelog.js and popup.js
-var imported = document.createElement("script");
-imported.src = "storelog.js";  
-document.getElementsByTagName("head")[0].appendChild(imported);
-imported.src = "popup.js";  
-document.getElementsByTagName("head")[0].appendChild(imported);
-
-// HTML elements that are used here:
-var previousGames = document.getElementById("previousGames");
-var message = document.getElementById('message');
-var backupFile = document.getElementById('backup-file');
-var ignoreBotBox = document.getElementById('ignoreBotGames');
-	
-	
 
 /**
  * function that clears local storage for all games that are stored. Cannot be 
@@ -20,13 +6,14 @@ var ignoreBotBox = document.getElementById('ignoreBotGames');
  * Also clears select boxes and the temporarily stored log.
  */
 function clearAllLogs(){
+	let previousGames = document.getElementById("previousGames");
 	if(confirm("Are you sure? This will delete all logs permanently from this computer")){
 		chrome.storage.local.get(["settings"], function(settings){
 			chrome.storage.local.clear(function(){
 				while (previousGames.options.length) {
 					previousGames.remove(0);
 				}
-				var value = {}
+				let value = {}
 				value["settings"] = settings;
 				chrome.storage.local.set(value, function(){
 					console.log("cleared logs",settings)
@@ -34,8 +21,7 @@ function clearAllLogs(){
 			});
 		});
 		
-		//Fill message field with text:
-		message.innerHTML = emptyMessage;
+		document.getElementById('LogPanel').innerHTML = emptyMessage;
 		
 		// send message to background script, to clear the temporarily stored log as well 
 		chrome.runtime.sendMessage({
@@ -54,16 +40,17 @@ function clearAllLogs(){
  * The date of the match is kept and also Name abbreviations are changed. 
  */
 async function changeNames(){
-	var log = message.innerHTML;
+	var LogPanel = document.getElementById('LogPanel');
+	let log = LogPanel.innerHTML;
 	if(log.match(/#\d+/) != null){
-		var players = [];
-		var abbr = [];
-		var newAbbr = [];
+		let players = [];
+		let abbr = [];
+		let newAbbr = [];
 		// get current gameID from the presented log.
 		gameID = log.match(/#\d+/).toString()
 		
 		// Load the corresponding game 
-		var game = await loadStoredGameByGameID(gameID);
+		const game = await loadStoredGameByGameID(gameID);
 		for(let i=0; i<game.players.length; i++){
 			// get the corrected player names (display the original ones)
 			players[i] = prompt("Player "+(i+1), game.players[i]);
@@ -104,16 +91,16 @@ async function changeNames(){
 		//Exchange names and abbreviations in the log 
 		for(let i=0; i<game.players.length; i++){
 			
-			var re = new RegExp(game.players[i], 'g');
+			let re = new RegExp(game.players[i], 'g');
 			game.log = game.log.replace(re, players[i]);
 			re = new RegExp(">"+abbr[i]+"<", 'g');
 			game.log = game.log.replace(re, ">"+newAbbr[i]+"<");
 		}
 		// store the modified log to the local storage
-		await storeLogLocally(game.log, game.uuid, game.kingdom, game.date);
+		await storeLogLocally(game.log, game.uuid, game.gameStatus, game.VPs, game.kingdom, game.date);
 		
 		// update the log in the browser view
-		message.innerHTML = game.log;
+		LogPanel.innerHTML = game.log;
 		
 		// refresh the previousGames drop down menu
 		console.log("Filling Previous Matches")
@@ -133,15 +120,15 @@ async function changeNames(){
  */
 function backupLogs() {
 	// set the fileName of the backup file with the current date in YYMMDD
-	var fileName = "LogDog_backup_"+getDateYYMMDD("")+".json";
+	const fileName = "LogDog_backup_"+getDateYYMMDD("")+".json";
 	
 	//load all games and save them to the download folder
 	chrome.storage.local.get(null, function(games) { 
 		// Convert object to a Json string.
-		var result = JSON.stringify(games);
+		const result = JSON.stringify(games);
 
 		// Save the Json file
-		var url = 'data:application/json;base64,' + btoa(result);
+		const url = 'data:application/json;base64,' + btoa(result);
 		chrome.downloads.download({
 			url: url,
 			filename: fileName
@@ -156,14 +143,14 @@ function backupLogs() {
 function addIfAbsent(backupGame){
 	return new Promise((resolve) => {
 		chrome.storage.local.get(null, function(games) {
-			var allGames = []
+			let allGames = []
 			
 			for(uuid in games){
 				if(games[uuid].gameID == backupGame.gameID){
 					allGames.push(games[uuid])
 				}
 			}
-			var found = false;
+			let found = false;
 			for(game of allGames){
 				try{
 					// include logs of identical logs and backup logs that are longer than the original but otherwise the same.
@@ -182,44 +169,22 @@ function addIfAbsent(backupGame){
 	});
 }
 
-
 /**
- * Event listener that is triggered, when a file is selected for loading a backup.
- * The backup will override games that are already in the list and add all games 
- * that are not present. 
+ * functio to control tabs
  */
-backupFile.addEventListener('change', function(){
-	//get the file that was selected as a backup to load
-	var file = backupFile.files[0];
-	if (file) {
-		var reader = new FileReader();
-		reader.readAsText(file, "UTF-8");
-		reader.onload = async function (evt) {
-			// parse String to JSON
-			var newGames = {};
-			var backupGames = JSON.parse(evt.target.result);
-			for(uuid of Object.keys(backupGames)){
-				var backupGame = backupGames[uuid];
-				// See, if the game is already present.
-				var newGame = await addIfAbsent(backupGame);
-				if(newGame != null){
-					newGames[newGame.uuid] = newGame;
-				}
-			}
-			chrome.storage.local.set(newGames, function() {
-				console.log("loaded backup")
-				chrome.storage.local.get(null, fillPreviousMatches)
-			});
-		}
-		reader.onerror = function (evt) {
-			console.log("error reading file");
-		}
-		
-		// set the file input value to null so that the same backup could be loaded again (e.g. if clearAllLogs was called after the backup.
-		backupFile.value = null;
+function clickHandler(elem) {
+	const target = elem.target;
+	const panelId = target.getAttribute('aria-controls'),
+		panel = document.getElementById(panelId);
+	if(panel!=null){
+		const selectedTab = document.querySelector('[aria-selected="true"]');
+		selectedTab.setAttribute('aria-selected', false);
+		target.setAttribute('aria-selected', true);
+		const panels = document.querySelector('[aria-hidden="false"]');
+		panels.setAttribute('aria-hidden', true);
+		panel.setAttribute('aria-hidden', false);
 	}
-});
-
+}
 
 /**
  * Install event handlers, when content is loaded
@@ -227,16 +192,60 @@ backupFile.addEventListener('change', function(){
 document.addEventListener('DOMContentLoaded', function() {
 	
 	//Event handlers
-
+	let previousGames = document.getElementById("previousGames");
+		
 	document.getElementById('clearLogs').addEventListener('click',clearAllLogs);
 	document.getElementById('backupLogs').addEventListener('click',backupLogs);
 	document.getElementById('changeNames').addEventListener('click',changeNames);
 
 	document.getElementById('loadBackup').addEventListener('click',function(){
-		backupFile.click();
+		document.getElementById('backup-file').click();
 	});
 	
+	document.getElementById('tablist').addEventListener('click',clickHandler);
+	
+	/**
+	 * Event listener that is triggered, when a file is selected for loading a backup.
+	 * The backup will override games that are already in the list and add all games 
+	 * that are not present. 
+	 */
+	 
+	let backupFile = document.getElementById('backup-file');
+	backupFile.addEventListener('change', function(){
+		//get the file that was selected as a backup to load
+		let file = backupFile.files[0];
+		if (file) {
+			let reader = new FileReader();
+			reader.readAsText(file, "UTF-8");
+			reader.onload = async function (evt) {
+				// parse String to JSON
+				let newGames = {};
+				let backupGames = JSON.parse(evt.target.result);
+				for(uuid of Object.keys(backupGames)){
+					let backupGame = backupGames[uuid];
+					// See, if the game is already present.
+					let newGame = await addIfAbsent(backupGame);
+					if(newGame != null){
+						newGames[newGame.uuid] = newGame;
+					}
+				}
+				chrome.storage.local.set(newGames, function() {
+					console.log("loaded backup")
+					chrome.storage.local.get(null, fillPreviousMatches)
+				});
+			}
+			reader.onerror = function (evt) {
+				console.log("error reading file");
+			}
+			
+			// set the file input value to null so that the same backup could be loaded again (e.g. if clearAllLogs was called after the backup.
+			backupFile.value = null;
+		}
+	});
+
+	
 	// Status of settings: needs to load getSetting and setSetting first (avoid bug by loading it a bit later)
+	let ignoreBotBox = document.getElementById('ignoreBotGames');
 	setTimeout(async function(){
 		ignoreBotBox.checked = await getSetting("ignoreBotGames", defaultValue = false);
 		ignoreBotBox.addEventListener('change',function(){