+++ popup.js
@@ -1,27 +1,31 @@
-// import functions from storelog.js
-var imported = document.createElement("script");
-imported.src = "storelog.js";  
-document.getElementsByTagName("head")[0].appendChild(imported);
-
 // Port that is opened to the background script.
 var myPort
+
+// close the image popup when clicking anywhere
+document.addEventListener("mousedown", function (evt) {
+	try{
+		div_hide();
+	}catch(e){}
+});
+// import functions from storelog.js
 
 /**
  * Downloads a single log as plain text to the download folder.
  */
 async function saveData(gameID) {
 	try{
+		let LogPanel = document.getElementById("LogPanel");
 		// Load the current game based on its gameID
-		var game = await loadStoredGameByGameID(gameID);
+		let game = await loadStoredGameByGameID(gameID);
 		
 		// Prepare log for saving and create fileName
-		message.innerHTML = game.log;
-		var log = message.innerText;
-		var date = game.date.replace(/\//g, "");
-		var fileName = ["LogDog",date,game.players[0],game.players[1],gameID+".txt"].join("_"); 
+		LogPanel.innerHTML = game.log;
+		let log = LogPanel.innerText;
+		let date = game.date.replace(/\//g, "");
+		let fileName = ["LogDog",date,game.players[0],game.players[1],gameID+".txt"].join("_"); 
 		
 		// parse game object to json
-		var json = JSON.stringify(log),
+		let json = JSON.stringify(log),
 			blob = new Blob([log], {type: "text/plain;charset=utf-8"}),
 			url = window.URL.createObjectURL(blob);
 		
@@ -39,10 +43,12 @@ async function saveData(gameID) {
 /**
  * Function that copies the currently viewed log to the clipboard
  */
-function copyTextToClipboard(text) {
+function copyTextToClipboard() {
+	let LogPanel = document.getElementById("LogPanel");
+	text = LogPanel.innerText;
 	if(text.length > 10){
 		//create an empty textArea to put in the text
-		var textArea = document.createElement("textarea");
+		let textArea = document.createElement("textarea");
 
 		// Position and style the textArea for minimal interference
 		textArea.style.position = 'fixed';
@@ -66,8 +72,8 @@ function copyTextToClipboard(text) {
 		
 		// copy the content of the textArea to the clipboard
 		try {
-		var successful = document.execCommand('copy');
-		var msg = successful ? 'Successful' : ' Not Successful';
+		let successful = document.execCommand('copy');
+		let msg = successful ? 'Successful' : ' Not Successful';
 			console.log('Copying text command was ' + msg);
 		} catch (e) {
 			console.log('Unable to copy to clipboard');
@@ -84,7 +90,7 @@ function copyTextToClipboard(text) {
  */
 function deleteSelectedLog(){
 	if(previousGames.options.length>0){
-		var delUUID = previousGames.options[previousGames.selectedIndex].value.split(",")[1];
+		let delUUID = previousGames.options[previousGames.selectedIndex].value.split(",")[1];
 		console.log(delUUID);
 		// delete the selected game from the storage and inform the background script to delete temporary data, if it is the last/current game that was deleted
 		chrome.storage.local.remove(delUUID, function(){
@@ -94,25 +100,91 @@ function deleteSelectedLog(){
 	}
 }
 
-
 /**
  * Load the last log that is in the selection box.
  */
-async function loadLastLog(){
+async function loadLog(index = -1){				
+	let previousGames = document.getElementById("previousGames");
+	let resultPanel = document.getElementById('ResultPanel');
+	let kingdomPanel = document.getElementById('KingdomPanel');
+	let LogPanel = document.getElementById("LogPanel");
 	
+	if (index == -1){
+		index = previousGames.options.length - 1
+	}
 	if(previousGames.options.length>0){
-		var lastGameID = previousGames.options[0].value.split(",")[0];
-		if(/#\d+/.test(lastGameID)){
-			var game = await loadStoredGameByGameID(lastGameID);
+		previousGames.selectedIndex = index
+		let gameID = previousGames.options[index].value.split(",")[0];
+		if(/#\d+/.test(gameID)){
+			let game = await loadStoredGameByGameID(gameID);
 			if(game!=undefined){
-				message.innerHTML = game.log;
+				let LogPanel = document.getElementById("LogPanel");
+				LogPanel.innerHTML = game.log;
+				getDateYYMMDD("")
+				if(kingdomPanel!= null){
+					kingdomPanel.innerText = "";
+					let txt = ""
+					try{
+						// Fill the log-tab and kingdom-tab 
+						for(VP of game.VPs){
+							txt=txt + VP.player + " (<span style='color:green;'><b>" + VP.vp_value + "VP</b></span>)<br> ";
+						}
+						let gameStatus = game.gameStatus;
+					
+						if(gameStatus!=undefined){
+							txt=txt+gameStatus;
+						}
+						
+						game.kingdom.kingdomCards.forEach(card =>{
+							let cardImage = new Image();
+							if(["Avanto", "Bustling Village", "Emporium", "Fortune", "Plunder", "Rocks"].indexOf(card) == -1){
+								cardImage.src = "images/dominionCards/"+card.replace(/ /g,"_")+"_small.jpg";
+								cardImage.addEventListener("click",function(a,b){
+									div_show(b,'images/dominionCards/'+card.replace(/ /g,"_")+'.jpg')
+								});
+								cardImage.style="margin:3px;";
+								kingdomPanel.appendChild(cardImage);
+							}
+
+						});
+						let licenseLink = document.createElement("div");
+						licenseLink.style="font-family:TrajanPro-Bold; font-letiant:small-caps; font-size:12pt;";
+						licenseLink.innerHTML = 'Images from <a href="http://wiki.dominionstrategy.com/">wiki.dominionstrategy.com</a> <br>under <a class="external" href="http://creativecommons.org/licenses/by-nc-sa/3.0/" style="user-select: auto;">Creative Commons Attribution Non-Commercial Share Alike</a> license'
+						kingdomPanel.appendChild(licenseLink);
+					}catch(e){kingdomPanel.innerText="Only available for new games (since LogDog 0.0.13)"}
+					resultPanel.innerHTML=txt
+				}
 			}
 			return;
 		}
 	}
 	console.log("no Log available");
-	message.innerHTML = "no Log available &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;";
+	LogPanel.innerHTML = "no Log available &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;";
 }
+
+
+/**
+ * function that shows shows full-size images of kingdom-card thumb-nails
+ */
+function div_show(ee,img){
+	div_hide();
+	let curImage= document.getElementById('currentImg');
+	curImage.src=img;
+	curImage.alt=img;
+	curImage.title=img;document.getElementById('largeImage').style.visibility='visible';
+	//set absolute position of full-size image
+	document.getElementById('largeImage').style.left=160+"px";
+	document.getElementById('largeImage').style.top=200+"px";
+}
+
+
+/**
+ * Function that hides the full-soze image.
+ */
+function div_hide(){
+	document.getElementById('largeImage').style.visibility='hidden';
+}
+
 
 /** 
  * function that sorts select boxes
@@ -122,8 +194,8 @@ async function loadLastLog(){
  
 function sortSelect(selElem, skip=0) {
     if(skip>0)skip--;
-	var tmpAry = new Array();
-    for (var i=skip;i<selElem.options.length;i++) {
+	let tmpAry = new Array();
+    for (let i=skip;i<selElem.options.length;i++) {
         tmpAry[i-skip] = new Array();
         tmpAry[i-skip][0] = selElem.options[i].text;
         tmpAry[i-skip][1] = selElem.options[i].value;
@@ -134,8 +206,8 @@ function sortSelect(selElem, skip=0) {
 	while (selElem.options.length > skip+1) {
         selElem.options[skip+1] = null;
     }
-    for (var i=0;i<tmpAry.length;i++) {
-        var op = new Option(tmpAry[i][0], tmpAry[i][1]);
+    for (let i=0;i<tmpAry.length;i++) {
+        let op = new Option(tmpAry[i][0], tmpAry[i][1]);
         selElem.options[i+skip] = op;
     }
     return;
@@ -148,11 +220,15 @@ function sortSelect(selElem, skip=0) {
  *               were applied
  */
 async function fillPreviousMatches(games, filteredBy=["any","any"]) {
+	let previousGames = document.getElementById("previousGames");
+	let playerSelect = document.getElementById('playerSelect');
+	let dateSelect = document.getElementById('dateSelect');
+	let LogPanel = document.getElementById("LogPanel");
 	
 	// get all game ids from the array of games
-	var allUUIDs = Object.keys(games);
+	let allUUIDs = Object.keys(games);
 	
-	var allGameIDs = [];
+	let allGameIDs = [];
 	for(uuid of allUUIDs){
 		if (checkUUID(uuid)){
 			allGameIDs.push(games[uuid].gameID);
@@ -161,44 +237,44 @@ async function fillPreviousMatches(games, filteredBy=["any","any"]) {
 			if(uuid.match(/#\d+/) != null){
 				// uuid is actually a gameID in the previous version
 				// create a new uuid
-				var oldGame = games[uuid];
-				var dateArray = oldGame.date.split("/");
-				var date = dateArray[2].substring(2)+dateArray[1]+dateArray[0];
-				var newID = getDateUUID(date);
-				var newGame = {
+				let oldGame = games[uuid];
+				let dateArray = oldGame.date.split("/");
+				let date = dateArray[2].substring(2)+dateArray[1]+dateArray[0];
+				let newID = getDateUUID(date);
+				let newGame = {
 					uuid: newID,
 					gameID: uuid,
 					players: [oldGame.player1,oldGame.player2],
 					date: oldGame.date,
 					kingdom: "",
+					VPs: "",
 					log: oldGame.log
 				};
 				allUUIDs.push(newID);
 				games[newID] = newGame;
-				await storeLogLocally(oldGame.log, newID, kingdom = "", dateString = oldGame.date, ignoreBotGames = false)
+				await storeLogLocally(oldGame.log, newID, gameStatus = "unknown", VPs = "", kingdom = "", dateString = oldGame.date, ignoreBotGames = false)
 				chrome.storage.local.remove(uuid);
 			}
 		}
 	}
-	
 	// first, remove all entries in the select box
 	for(i = previousGames.options.length-1; i>=0; i--){previousGames.remove(i);}
 	
 	// create arrays to hold all available players and dates that occur among all games
-	var players = [];
-	var dates = [];
+	let players = [];
+	let dates = [];
 	
 	allUUIDs.sort();
 	// get all players and dates of all games and put the games in the selection box
 	for(uuid of allUUIDs){
 		if (checkUUID(uuid)){
-			var game = games[uuid];
+			let game = games[uuid];
 			console.log()
-			var gameID = game.gameID;
+			let gameID = game.gameID;
 			players.push(game.players[0]);
 			players.push(game.players[1]);
 			dates.push(game.date);
-			var el = document.createElement("option");
+			let el = document.createElement("option");
 			el.textContent = [gameID, game.players.join(" vs. "),"(" + game.date + ")"].join(" ");;
 			el.value = [gameID, uuid].join(",");
 			previousGames.appendChild(el);
@@ -218,19 +294,19 @@ async function fillPreviousMatches(games, filteredBy=["any","any"]) {
 			for(i = dateSelect.options.length-1; i>=0; i--){dateSelect.remove(i);}
 			
 			// Add default fields "Filter for Date" and "Any" to the drop down
-			var el = document.createElement("option");
+			let el = document.createElement("option");
 			el.textContent = "Filter for Date";
 			el.value = "any";
 			el.style="display:none;";
 			dateSelect.appendChild(el);
-			var el = document.createElement("option");
+			el = document.createElement("option");
 			el.textContent = "Any";
 			el.value = "any";
 			dateSelect.appendChild(el);
 			
 			// Add all unique dates to the drop down.
 			for(date of dates){
-				var el = document.createElement("option");
+				let el = document.createElement("option");
 				el.textContent = date;
 				el.value = date;
 				dateSelect.appendChild(el);
@@ -249,19 +325,19 @@ async function fillPreviousMatches(games, filteredBy=["any","any"]) {
 			for(i = playerSelect.options.length-1; i>=0; i--){playerSelect.remove(i);}
 			
 			// Add default fields "Filter for Player" and "Any" to the drop down
-			var el = document.createElement("option");
+			let el = document.createElement("option");
 			el.textContent = "Filter for Player";
 			el.value = "any";
 			el.style="display:none;";
 			playerSelect.appendChild(el);
-			var el = document.createElement("option");
+			el = document.createElement("option");
 			el.textContent = "Any";
 			el.value = "any";
 			playerSelect.appendChild(el);
 			
 			// Add all unique players to the drop down.
 			for(player of players){
-				var el = document.createElement("option");
+				let el = document.createElement("option");
 				el.textContent = player;
 				el.value = player;
 				playerSelect.appendChild(el);
@@ -285,9 +361,9 @@ async function fillPreviousMatches(games, filteredBy=["any","any"]) {
 	}// only works in options.html
 
 	try{
-		loadLastLog();
+		loadLog();
 	}catch(e){
-		message.innerHTML=emptyMessage;
+		LogPanel.innerHTML=emptyMessage;
 		console.log("Error loading last log");
 		console.log(e);
 	}
@@ -301,10 +377,14 @@ async function fillPreviousMatches(games, filteredBy=["any","any"]) {
  * selection.
  */
 function filterGames(){
+	let playerSelect = document.getElementById('playerSelect');
+	let dateSelect = document.getElementById('dateSelect');
+	let LogPanel = document.getElementById("LogPanel");
 	
 	//get the values that will be used for filtering ("any" is one of the options).
-	var player = playerSelect.options[playerSelect.selectedIndex].value
-	var date = dateSelect.options[dateSelect.selectedIndex].value
+	let player = playerSelect.options[playerSelect.selectedIndex].value
+	let date = dateSelect.options[dateSelect.selectedIndex].value
+	let previousGames = document.getElementById("previousGames");
 	
 	
 	// remove all games from the drop down menu first
@@ -313,13 +393,13 @@ function filterGames(){
 	// Load all games and filter for the matching ones 
 	chrome.storage.local.get(null,async function(games){
 		// get all game IDs
-		var allUUIDs = Object.keys(games);
+		let allUUIDs = Object.keys(games);
 		// create a new array that will hold the filtered game ids
-		var filteredGames = {};
+		let filteredGames = {};
 		for(uuid of allUUIDs){
 			//check if the uuid key actually is a uuid to prevent exceptions
 			if (checkUUID(uuid)){
-				var game = games[uuid];
+				let game = games[uuid];
 				//compare the filter to each game and push matching ids into gameIDs
 				if((date.trim() == "any" || game.date == date)&&(player.trim() == "any" || game.players.indexOf(player) != -1)){
 					filteredGames[uuid] = game;
@@ -332,9 +412,9 @@ function filterGames(){
 		
 		// Load the last entry of the list
 		try{
-			loadLastLog();
+			loadLog();
 		}catch(e){
-			message.innerHTML=emptyMessage;
+			LogPanel.innerHTML=emptyMessage;
 			console.log("error loading log, probably no log file saved");
 			console.log(e);
 		}
@@ -349,21 +429,12 @@ function filterGames(){
  * All event listeners are setup in here, as well as the port to the background 
  * script
  */	
-function onWindowLoad() {
+document.addEventListener('DOMContentLoaded', function() {
 	try{
-		// HTML elements that are used here:
-		var message = document.getElementById("message");
-		var previousGames = document.getElementById("previousGames");
-		var downloadBtn = document.getElementById("downloadLog");
-		var copyBtn = document.getElementById("copyLog");
-		var dateSelect = document.getElementById("dateSelect");
-		var playerSelect = document.getElementById("playerSelect");
-		var optionsButton = document.getElementById("options");
-		var loadGameButton = document.getElementById("loadGame");
-		var loadKingdomButton = document.getElementById("loadKingdom");
-		var playerSelect = document.getElementById('playerSelect');
-		var dateSelect = document.getElementById('dateSelect');
-
+		let previousGames = document.getElementById("previousGames");
+		let LogPanel = document.getElementById("LogPanel");
+		
+		
 		//Load all previous games into the popup option selector 
 		chrome.storage.local.get(null, fillPreviousMatches);
 		
@@ -380,13 +451,14 @@ function onWindowLoad() {
 			// retrieve the log from the message
 			log = m.log;
 			if(log != null && log.length > 0){
-				// view the log in the message div of the popup/options window
-				message.innerHTML = log;
+				
+				// view the log in the LogPanel div of the popup/options window
+				LogPanel.innerHTML = log;
 				// retrieve the information whether the game has finished
-				var game_over = m.timeout;
+				let gameStatus = m.gameStatus;
 				
 				// retrieve the game id from the log
-				var gameID = "";
+				let gameID = "";
 				try{
 					gameID = log.match(/#\d+/).toString();
 				}catch(e){
@@ -395,8 +467,8 @@ function onWindowLoad() {
 				}
 				
 				// Select the current log in the drop down or make sure that it is not present in the list (found remains false)
-				var found = false;
-				var value = [gameID, m.uuid].join(",");
+				let found = false;
+				let value = [gameID, m.uuid].join(",");
 				for(i = 0; i< previousGames.length; i++){
 					if(value === previousGames[i].value){
 						previousGames.selectedIndex = i;
@@ -407,64 +479,58 @@ function onWindowLoad() {
 				
 				// If the log is not yet listed, add it to the list of previous games and and save it to storage.local
 				if(!found){
+					
 					if(/#\d+/.test(gameID)){
 						
 						// Store the game to storage.local 
-						storeLogLocally(log, m.uuid, kingdom = m.kingdom);
+						storeLogLocally(log, m.uuid, gameStatus = m.gameStatus, VPs = m.VPs, kingdom = m.kingdom);
 						
 						// get the game object
-						var game = await loadStoredGameByGameID(gameID);
+						let game = await loadStoredGameByGameID(gameID);
 												
 						// add the game to the list of previous games
-						var el = document.createElement("option");
+						let el = document.createElement("option");
 						el.textContent = [gameID, game.players[0],"vs.", game.players[1],"(" + game.date + ")"].join(" ");;
 						
 		
 						el.value = [gameID, uuid].join(",");
 		
 						previousGames.appendChild(el);
-						
 						// select this game entry
 						previousGames.selectedIndex = previousGames.length - 1;
 					}
 				}
 				// If the log is faulty (too short or mismatching gameID), load the last log
 				if(log.length<10){
-					loadLastLog();
+					loadLog();
 				}
 			}
 		});
 		
-		// Event listener that will load the selected game from the drop down menu into the message div
-		previousGames.addEventListener("change", async function(){
-			
-			// get the ID of the selected game
-			var value = previousGames.options[previousGames.selectedIndex].value;
-			
-			// load the selected game
-			var game = await loadStoredGameByUUID(value.split(",")[1]);
-			
-			// view the log 
-			message.innerHTML = game.log;
+		// Event listener that will load the selected game from the drop down menu into the LogPanel div
+		previousGames.addEventListener("change", function(){
+			loadLog(previousGames.selectedIndex);
 		});
 		
 		
 		// Button event handlers, calling above functions.
 		
-		playerSelect.addEventListener("change", filterGames);
-		dateSelect.addEventListener("change", filterGames);
+		
+		document.getElementById('playerSelect').addEventListener("change", filterGames);
+		document.getElementById('dateSelect').addEventListener("change", filterGames);
 	
 		document.getElementById('downloadLog').addEventListener('click',function(){
-			saveData(message.innerText.match(/#\d+/).toString());
+			saveData(LogPanel.innerText.match(/#\d+/).toString());
 		});
 		
-		document.getElementById('copyLog').addEventListener('click',function(){copyTextToClipboard(message.innerText)});
+		document.getElementById('copyLog').addEventListener('click',function(){copyTextToClipboard()});
 		document.getElementById('deleteLog').addEventListener('click',function(){deleteSelectedLog()});
 		
+		let loadKingdomButton = document.getElementById("loadKingdom");
 		// Load a new game with kingdom cards of the selected game
 		if(loadKingdomButton!=null){
 			loadKingdomButton.addEventListener('click',async function(){
-				var game = await loadStoredGameByGameID(previousGames.options[previousGames.selectedIndex].value.split(",")[0]);
+				let game = await loadStoredGameByGameID(previousGames.options[previousGames.selectedIndex].value.split(",")[0]);
 				chrome.tabs.query({ active: true, currentWindow: true }, tabs => {
 					try{
 						let tab = tabs[0];
@@ -484,9 +550,11 @@ function onWindowLoad() {
 			});
 		}
 		
+		
+		let loadGameButton = document.getElementById("loadGame");
 		if(loadGameButton!=null){
 			loadGameButton.addEventListener('click', function(){
-				var gameID = previousGames.options[previousGames.selectedIndex].value.split(",")[0];
+				let gameID = previousGames.options[previousGames.selectedIndex].value.split(",")[0];
 				chrome.tabs.query({ active: true, currentWindow: true }, tabs => {
 					try{
 						let tab = tabs[0];
@@ -505,6 +573,8 @@ function onWindowLoad() {
 				});
 			});
 		}
+		
+		let optionsButton = document.getElementById("options");
 		if(optionsButton!=null){
 			optionsButton.addEventListener('click',function() {
 				if (chrome.runtime.openOptionsPage) {
@@ -518,7 +588,7 @@ function onWindowLoad() {
 		console.log("Error while loading the popup");
 		console.log(e);
 	}
-}
+})
 
-window.onload = onWindowLoad;
+